Moved readme parser to folder
